+++ src/lib/autostart.ts
@@ -1,0 +1,23 @@
+import { enable, isEnabled, disable } from "tauri-plugin-autostart-api";
+
+export const enableAutostart = async () => {
+  if (await isEnabled()) {
+    return;
+  }
+
+  await enable();
+};
+
+export const disableAutostart = async () => {
+  if (!(await isEnabled())) {
+    return;
+  }
+
+  await disable();
+};
+
+export const autostart = {
+  enable: enableAutostart,
+  disable: disableAutostart,
+  isEnabled,
+};