+++ src/routes/index.tsx
@@ -3,6 +3,7 @@ import Layout from "@/layout";
 import DashboardPage from "@/features/metrics/pages/dashboard.page";
 import DisksPage from "@/features/metrics/pages/disks.page";
 import ProcessesPage from "@/features/processes/pages/processes.page";
+import SettingsPage from "@/features/settings/pages/settings.page";
 
 const AppRoutes: React.FC = () => {
   return (
@@ -12,6 +13,7 @@ const AppRoutes: React.FC = () => {
           <Route path="/" element={<DashboardPage />} />
           <Route path="/disks" element={<DisksPage />} />
           <Route path="/processes" element={<ProcessesPage />} />
+          <Route path="/settings" element={<SettingsPage />} />
         </Route>
       </Routes>
     </Router>